refactor: move URL utilities from runtime/plan_resolver.rs to data/url.rs
The file no longer contains the PlanResolver trait (moved to
vegafusion-runtime in an earlier commit). Rename to reflect its
actual contents: URL types, resolution, and parsing utilities.
