+++ vegafusion-core/src/data/mod.rs
@@ -1,2 +1,3 @@
 pub mod dataset;
 pub mod tasks;
+pub mod url;
