+++ vegafusion-core/src/runtime/mod.rs
@@ -1,7 +1,6 @@
-mod plan_resolver;
 mod runtime;
 
-pub use plan_resolver::{
+pub use crate::data::url::{
     has_url_scheme, is_absolute_path, normalize_base_url, path_to_file_url, resolve_data_base_url,
     resolve_url, DataBaseUrlSetting, ParsedUrl, ResolutionResult,
 };
